+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Improve error reporting in prometheus scaler ([#1497](https://github.com/kedacore/keda/pull/1497))
 - Check that metricNames are unique in ScaledObject ([#1390](https://github.com/kedacore/keda/pull/1390))
 - Serve OpenAPI spec from KEDA Metrics Apiserver ([#1512](https://github.com/kedacore/keda/pull/1512))
+- Support metrics with multiple dimensions and configurable metricValues on AWS Cloudwatch Scaler ([#1230](https://github.com/kedacore/keda/issues/1230))
 
 ### Breaking Changes